abstract out createLocalScope
