+++ src/flowCoffeescriptKernel/flowCoffeescriptKernel.js
@@ -4,6 +4,7 @@ import { parseJavascript } from './parseJavascript';
 import { parseDeclarations } from './parseDeclarations';
 import { traverseJavascript } from './traverseJavascript';
 import { deleteAstNode } from './deleteAstNode';
+import { createLocalScope } from './createLocalScope';
 
 export function flowCoffeescriptKernel() {
   const lodash = window._;
@@ -11,27 +12,6 @@ export function flowCoffeescriptKernel() {
   const escodegen = window.escodegen;
   const esprima = window.esprima;
   const CoffeeScript = window.CoffeeScript;
-  const createLocalScope = node => {
-    let param;
-    let _i;
-    let _len;
-    // parse all declarations in this scope
-    const localScope = parseDeclarations(node.body);
-
-    // include formal parameters
-    const _ref = node.params;
-    for (_i = 0, _len = _ref.length; _i < _len; _i++) {
-      param = _ref[_i];
-      if (param.type === 'Identifier') {
-        localScope[param.name] = {
-          name: param.name,
-          object: 'local',
-        };
-      }
-    }
-    return localScope;
-  };
-
   // redefine scope by coalescing down to non-local identifiers
   const coalesceScopes = scopes => {
     let i;
